Update ApiType base class
This updates the ApiType base class that all of our value objects are
based on. We were defining the class using strict mode, which was a
problem because GraphQL lazilt loads data. For instance when loading all
users, we can load all of the todo lists at the same time, but we cannot
load all of the todos, because GraphQL has not traversed that far down
in the query yet.

The initial solution to this problem is to change the constructor to be
a `strict_with_default` constructor, that allows us to maintain strict
type safety, and initialize the object with a default.

It would be great in the future to be able to be smarter about our
queries.

This may be useful: https://github.com/shopify/graphql-batch